Fixed weapon pickup/drop, fixed sniper damage,
sniper damage was only working if the enemies health was less then the sniper damage.

moved the pickup/drop funtions to the player so that the only entity looking for input at any given time was the player and then the weapons called the functions that the input was used.